add register page

+++ routes/smartlist.js
@@ -4,6 +4,12 @@ const router  = express.Router();
 module.exports = (db) => {
   // get request for loading page (all tasks)
   router.get("/", (req, res) => {
+
+    if (!req.cookies["user_id"]) {
+      res.status(401).send("You need to log in to do that!");
+      res.redirect('/');
+    }
+
     let query = `SELECT todos.* FROM todos`;
     db.query(query)
       .then(data => {
@@ -20,20 +26,26 @@ module.exports = (db) => {
 
   // post request to add a task (all data for todos schema)
   router.post("/", (req, res) => {
+
+    if (!req.cookies["user_id"]) {
+      res.status(401).send("You need to log in to do that!");
+      return;
+    }
+
     // API REQUEST WITH DATA
     // USE TASK req.body.task to go through API
     const taskName = req.body.task.toLowerCase();
     let category_id;
 
     // pre-emptive sorting for certain words
-    if (taskName.includes('watch')) {
-      category_id = 1;
+    if (taskName.includes('buy')) {
+      category_id = 4;
     } else if (taskName.includes('eat')) {
       category_id = 2;
     } else if (taskName.includes('read')) {
       category_id = 3;
-    } else if (taskName.includes('buy')) {
-      category_id = 4;
+    } else if (taskName.includes('watch')) {
+      category_id = 1;
     } else {
       category_id = 5;
     }
@@ -46,11 +58,11 @@ module.exports = (db) => {
     }
 
     let query = `INSERT INTO todos (task, category_id, user_id, important_tasks) VALUES ('${todos.task}', ${todos.category_id}, ${todos.user_id}, ${todos.important_tasks}) RETURNING *;`;
-    console.log(query);
+
     db.query(query)
       .then(data => {
-
         const todo = data.rows[0];
+        console.log(todo)
         res.json({ todo });
       })
       .catch(err => {
@@ -62,6 +74,12 @@ module.exports = (db) => {
 
   // delete request to delete a task (all data for todos schema)
   router.delete("/:id", (req, res) => {
+
+    if (!req.cookies["user_id"]) {
+      res.status(401).send("You need to log in to do that!");
+      return;
+    }
+
     let deleteId = req.params.id;
     // console.log(deleteId)
     let query = `DELETE FROM todos WHERE id = ${deleteId}`;
